Add validator for Cabinet Office emails
This is a non-exhaustive and extensible approach to verifying Cabinet Office email. There is more than one domain within the CO, and this code attempts to be flexible enough to deal with all of them.

There is no validation on whether the user accidentally added a space or comma or similar - they will need to try it again.